+++ src/seg_sites.cpp
@@ -3,41 +3,15 @@
 using namespace Rcpp;
 
 
-//' Segregating Sites
-//'
-//' This functions allow the creation and modification of segregating sites
-//' objects, which are one of the most basic intermediary statistics that is
-//' calculated in coala. Segregating sites are primarily SNP matrix that
-//' contains all SNPs for one locus, with some additional information attached.
-//' The parts of the S3 class are detailed below.
-//'
-//' A segregating sites object contains all SNPs for one genetic locus. Each
-//' object consists of tree parts: A SNP matrix, a vector of SNP positons and
-//' a vector that states which transcript a SNP belong to, if the locus
-//' consists of multiple transscripts ('locus trio').
-//'
-//' \itemize{
-//'   \item{In the \strong{SNP} matrix, each row represents a haplotype and each
-//'         column represents a SNP. An entry is either \code{1} if the
-//'         haplotype carries the derived allele for the SNP, or \code{0} if it
-//'         carries the ancestral one.}
-//'   \item{In the \strong{positions} vector, each entry gives the relative
-//'         position of SNP in the corresponding column of the SNP matrix.}
-//'   \item{The \strong{trio_locus} vector contains the trio locus each SNP
-//'         belongs to. Entry of \code{-1},\code{0}, \code{1} represent the
-//'         left, middle, and right locus, respectively. For normal loci,
-//'         this just consists of \code{0}'s}
-//' }
+//' @describeIn segsites Creates segregating sites
 //'
 //' @param snps The SNP Matrix (see Details).
 //' @param positions A numeric vector indicating the relative positions of each
 //'   SNP on the locus (see Details).
 //' @param trio_locus If the locus consists of a locus trio (see Details).
-//' @param check Whether non-segregating sites are remove from the segregating
+//' @param check Whether non-segregating sites are removed from the segregating
 //'   sites (\code{TRUE}) or not (\code{FALSE}).
 //' @export
-//' @aliases segsites
-//' @author Paul Staab
 //'
 // [[Rcpp::export]]
 coala::SegSites create_segsites(NumericMatrix snps,
@@ -48,7 +22,7 @@ coala::SegSites create_segsites(NumericMatrix snps,
 }
 
 
-//' @describeIn create_segsites Returns the SNP matrix from a segregating sites
+//' @describeIn segsites Returns the SNP matrix from a segregating sites
 //'  object.
 //'
 //' @param segsites The segregating sites object
@@ -59,7 +33,7 @@ NumericMatrix get_snps(const coala::SegSites segsites) {
 }
 
 
-//' @describeIn create_segsites Returns the SNP's positions from a segregating
+//' @describeIn segsites Returns the SNP's positions from a segregating
 //'   sites  object.
 //' @export
 // [[Rcpp::export]]
@@ -68,7 +42,7 @@ NumericVector get_positions(const coala::SegSites segsites) {
 }
 
 
-//' @describeIn create_segsites Sets the SNP's positions in a segregating
+//' @describeIn segsites Sets the SNP's positions in a segregating
 //'   sites object.
 //' @export
 // [[Rcpp::export]]
@@ -78,7 +52,7 @@ coala::SegSites set_positions(coala::SegSites segsites,
 }
 
 
-//' @describeIn create_segsites Returns the trio locus positions from a
+//' @describeIn segsites Returns the trio locus positions from a
 //'   segregating sites  object.
 //' @export
 // [[Rcpp::export]]
@@ -87,7 +61,7 @@ NumericVector get_trio_locus(const coala::SegSites segsites) {
 }
 
 
-//' @describeIn create_segsites Sets the trio locus in a segregating sites
+//' @describeIn segsites Sets the trio locus in a segregating sites
 //'   object.
 //' @export
 // [[Rcpp::export]]