+++ tests/testthat/test-model-class.R
@@ -129,33 +129,33 @@ test_that("getting the available Populations works", {
 
 
 test_that("get population individuals works", {
-  expect_equal(get_population_indiviuals(model_theta_tau(), 1), 1:10)
-  expect_equal(get_population_indiviuals(model_theta_tau(), 2), 11:25)
-  expect_equal(get_population_indiviuals(model_theta_tau(), "all"), 1:25)
-  expect_error(get_population_indiviuals(model_theta_tau(), 3))
-  expect_error(get_population_indiviuals(model_theta_tau(), "al"))
+  expect_equal(get_population_individuals(model_theta_tau(), 1), 1:10)
+  expect_equal(get_population_individuals(model_theta_tau(), 2), 11:25)
+  expect_equal(get_population_individuals(model_theta_tau(), "all"), 1:25)
+  expect_error(get_population_individuals(model_theta_tau(), 3))
+  expect_error(get_population_individuals(model_theta_tau(), "al"))
 
   # With an outgroup
-  expect_equal(get_population_indiviuals(model_hky(), "all"), 1:6)
-  expect_equal(get_population_indiviuals(model_hky(), 1), 1:3)
-  expect_equal(get_population_indiviuals(model_hky(), 2), 4:6)
-  expect_error(get_population_indiviuals(model_hky(), 3))
+  expect_equal(get_population_individuals(model_hky(), "all"), 1:6)
+  expect_equal(get_population_individuals(model_hky(), 1), 1:3)
+  expect_equal(get_population_individuals(model_hky(), 2), 4:6)
+  expect_error(get_population_individuals(model_hky(), 3))
 
   model <- coal_model(1:3) + feat_outgroup(2)
-  expect_equal(get_population_indiviuals(model, "all"), 1:4)
-  expect_equal(get_population_indiviuals(model, 1), 1)
-  expect_error(get_population_indiviuals(model, 2))
-  expect_equal(get_population_indiviuals(model, 3), 2:4)
+  expect_equal(get_population_individuals(model, "all"), 1:4)
+  expect_equal(get_population_individuals(model, 1), 1)
+  expect_error(get_population_individuals(model, 2))
+  expect_equal(get_population_individuals(model, 3), 2:4)
 })
 
 
 test_that("getting indiviuals in polyploid models works", {
   model <- coal_model(1:5, ploidy = 2)
-  expect_equal(get_population_indiviuals(model, 1, haploids = FALSE), 1)
-  expect_equal(get_population_indiviuals(model, 2, haploids = FALSE), 2:3)
-  expect_equal(get_population_indiviuals(model, 3, haploids = FALSE), 4:6)
-  expect_equal(get_population_indiviuals(model, 4, haploids = FALSE), 7:10)
-  expect_equal(get_population_indiviuals(model, 5, haploids = FALSE), 11:15)
+  expect_equal(get_population_individuals(model, 1, haploids = FALSE), 1)
+  expect_equal(get_population_individuals(model, 2, haploids = FALSE), 2:3)
+  expect_equal(get_population_individuals(model, 3, haploids = FALSE), 4:6)
+  expect_equal(get_population_individuals(model, 4, haploids = FALSE), 7:10)
+  expect_equal(get_population_individuals(model, 5, haploids = FALSE), 11:15)
 })
 
 
@@ -252,4 +252,72 @@ test_that("model checking give not errors", {
   capture.output(check_model(model_gtr()))
   capture.output(check_model(model_hky()))
   capture.output(check_model(model_trios()))
+})
+
+
+test_that("model parts can be combined into a partial model", {
+  incomplete_model <- feat_growth(1, 1) + feat_growth(2, 2)
+  expect_true(is_partial_model(incomplete_model))
+  expect_equal(length(incomplete_model), 2)
+
+  incomplete_model <- sumstat_sfs() + sumstat_dna()
+  expect_true(is_partial_model(incomplete_model))
+  expect_equal(length(incomplete_model), 2)
+
+  incomplete_model <- par_const(5) + par_const(7)
+  expect_true(is_partial_model(incomplete_model))
+  expect_equal(length(incomplete_model), 2)
+
+  incomplete_model <- locus_single(1) + locus_averaged(2, 10)
+  expect_true(is_partial_model(incomplete_model))
+  expect_equal(length(incomplete_model), 2)
+})
+
+
+test_that("partial models can be extended", {
+  incomplete_model <- feat_growth(1, 1) +
+    feat_mutation(5) +
+    feat_recombination(7)
+  expect_true(is_partial_model(incomplete_model))
+  expect_equal(length(incomplete_model), 3)
+  expect_equal(incomplete_model[[1]], feat_growth(1, 1))
+  expect_equal(incomplete_model[[2]], feat_mutation(5))
+  expect_equal(incomplete_model[[3]], feat_recombination(7))
+
+  incomplete_model_2 <- incomplete_model + sumstat_sfs()
+  expect_true(is_partial_model(incomplete_model_2))
+  expect_equal(length(incomplete_model_2), 4)
+  expect_equal(incomplete_model_2[[4]], sumstat_sfs())
+})
+
+
+test_that("partial models can be added to models", {
+  incomplete_model <- feat_growth(1, 1) +
+    feat_mutation(5) +
+    feat_recombination(par_const(8)) +
+    locus_averaged(10, 100) +
+    par_const(6) +
+    sumstat_sfs()
+  model <- coal_model(10) + incomplete_model
+
+  model_direct <- coal_model(10) +
+    feat_growth(1, 1) +
+    feat_mutation(5) +
+    feat_recombination(par_const(8)) +
+    locus_averaged(10, 100) +
+    par_const(6) +
+    sumstat_sfs()
+  model_direct$id <- model$id
+
+  expect_equal(model, model_direct)
+})
+
+
+test_that("printing partical models works", {
+  incomplete_model <- feat_growth(1, 1) +
+    feat_mutation(5) +
+    feat_recombination(par_const(8))
+  expect_output(print(incomplete_model), "growth")
+  expect_output(print(incomplete_model), "Mutation")
+  expect_output(print(incomplete_model), "Recombination")
 })